rewrite: merge dice.js into math.js, add clown cmd
